chore: Adjusting props locations and employee mocks values
