+++ src/components/button/button.tsx
@@ -1,5 +1,13 @@
-import { ButtonProps } from "../../interfaces/props";
 import styles from "./button.module.css";
+
+export interface ButtonProps {
+    label: string,
+    onClick?: () => void;
+    variant?: 'primary' | 'secondary' | 'error' | 'outlined' | 'outlined-secondary' | 'outlined-error' | 'outlined-neutral';
+    disabled?: boolean;
+    size?: 'ssm' | 'sm' | 'md' | 'lg' | 'xg';
+    type?: 'button' | 'submit' | 'reset';
+};
 
 function Button({
     label,
