+++ src/components/textInput/textInput.tsx
@@ -1,5 +1,14 @@
-import { TextInputProps } from '../../interfaces/props';
 import styles from './textInput.module.css';
+
+export interface TextInputProps {
+    placeholder: string;
+    size?: 'sm' | 'md' | 'lg' | 'xg';
+    variant?: 'default' | 'underline';
+    value?: string;
+    readonly?: boolean;
+    type?: 'text' | 'password';
+    onChange?: (value: string) => void;
+};
 
 function TextInput({
     placeholder,
