--- src/interfaces/props/buttonProps.ts
@@ -1,8 +1,0 @@
-export interface ButtonProps {
-    label: string,
-    onClick?: () => void;
-    variant?: 'primary' | 'secondary' | 'error' | 'outlined' | 'outlined-secondary' | 'outlined-error' | 'outlined-neutral';
-    disabled?: boolean;
-    size?: 'ssm' | 'sm' | 'md' | 'lg' | 'xg';
-    type?: 'button' | 'submit' | 'reset';
-};
