--- src/interfaces/props/employeeCardProps.ts
@@ -1,6 +1,0 @@
-import { Employee } from "../models";
-
-export interface EmployeeCardProps {
-    employee: Employee;
-    onClick: () => void;
-};
