--- src/interfaces/props/employeeDetailsProps.ts
@@ -1,8 +1,0 @@
-import { Employee } from "../models";
-
-export interface EmployeeDetailsProps {
-    employee: Employee;
-    isOpen: boolean;
-    onClose: () => void;
-    onRemove: () => void;
-};
