--- src/interfaces/props/headerProps.ts
@@ -1,5 +1,0 @@
-import React from "react";
-
-export interface HeaderProps {
-    children: React.ReactNode;
-};
