--- src/interfaces/props/index.ts
@@ -1,12 +1,0 @@
-export * from './buttonProps';
-export * from './headerProps';
-export * from './postCardProps';
-export * from './tagProps';
-export * from './textInputMultilineProps';
-export * from './textInputProps';
-export * from './switchProps';
-export * from './employeeCardProps';
-export * from './employeeDetailsProps';
-export * from './postDetailsProps';
-export * from './solicitationDetailsProps';
-export * from './solicitationCardProps';
