--- src/interfaces/props/postCardProps.ts
@@ -1,6 +1,0 @@
-import { Post } from "../models";
-
-export interface PostCardProps {
-    post: Post;
-    onClick?: () => void;
-};
