--- src/interfaces/props/postDetailsProps.ts
@@ -1,7 +1,0 @@
-import { Post } from "../models";
-
-export interface PostDetailsProps {
-    post: Post;
-    isOpen: boolean;
-    onClose: () => void;
-};
