--- src/interfaces/props/solicitationCardProps.ts
@@ -1,8 +1,0 @@
-import { Solicitation } from "../models";
-
-export interface SolicitationCardProps {
-    solicitation: Solicitation;
-    onClick: () => void;
-    onAccept: () => void;
-    onRefuse: () => void;
-};
