--- src/interfaces/props/solicitationDetailsProps.ts
@@ -1,9 +1,0 @@
-import { Solicitation } from "../models";
-
-export interface SolicitationDetailsProps {
-    solicitation: Solicitation;
-    isOpen: boolean;
-    onAccept: () => void;
-    onRefuse: () => void;
-    onClose: () => void;
-};
