--- src/interfaces/props/switchProps.ts
@@ -1,4 +1,0 @@
-export interface switchProps {
-    isOn?: boolean;
-    onChange?: (checked: boolean) => void;
-};
